Refactor repeated code
