+++ packages/grid/src/GridUtils.ts
@@ -1,5 +1,6 @@
 import React from 'react';
 import clamp from 'lodash.clamp';
+import { find as linkifyFind } from 'linkifyjs';
 import { EMPTY_ARRAY } from '@deephaven/utils';
 import GridRange, { GridRangeIndex } from './GridRange';
 import {
@@ -1438,6 +1439,22 @@ export class GridUtils {
       y2: newBottom,
       token: tokenBox.token,
     };
+  }
+
+  /**
+   * Finds tokens in text (urls, emails) that start with https:// or http://
+   * @param text The text to search in
+   * @returns An array of tokens
+   */
+  static findTokensWithProtocolInText(text: string): Token[] {
+    const tokens = linkifyFind(text);
+
+    return tokens.filter(token => {
+      if (token.type === 'url') {
+        return /^https?:\/\//.test(token.value);
+      }
+      return true;
+    });
   }
 }
